Fix simple compiler errors when building the Java Debugger agent's C++ code using Clang on OSX. Clang on OSX is a little more pedantic.
1. -Wunused-const-variable
Remove unused constant variables.

2. -Wmismatched-tags
For "struct Foo{ }" types, all forward declarations must also be "struct"

3. -Wtautological-constant-out-of-range-compare
Comparing a char to 0xC0 is always false. Must first cast string elements to uint8.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150921661

+++ src/agent/value_formatter.cc
@@ -49,7 +49,8 @@ static void ScrubEmbeddedZeroCharacters(string* data_ptr) {
   std::vector<int> zeros;
   for (int i = 0; i + 1 < data.size(); ++i) {
     // Check if the next two characters represent an embedded zero.
-    if (data[i] == 0xC0 && data[i + 1] == 0x80) {
+    if (static_cast<uint8>(data[i]) == 0xC0 &&
+        static_cast<uint8>(data[i + 1]) == 0x80) {
       zeros.push_back(i);
       ++i;  // Skip the next byte (0x80).
       continue;
